third_party/blink/web_tests/svg/dynamic-updates/SVGMaskElement-dom-height-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'height' attribute of the SVGMaskElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS maskElement.getAttribute('height') is "10%"
PASS maskElement.getAttribute('height') is "100%"
PASS successfullyParsed is true

TEST COMPLETE
